[FIX] booking engine calendar
