+++ pms_pwa/static/src/js/o_pms_pwa_reservation_table_filter.js
@@ -34,1152 +34,6 @@ odoo.define("pms_pwa.reservation_table", function (require) {
         }
         $("form").submit();
     });
-    /* Function form to json*/
-    function form_to_json(formData) {
-        var form_object = {};
-        $.each(formData, function (i, v) {
-            if (v.value != "false") {
-                if (v.name in form_object) {
-                    if(typeof form_object[v.name] == 'object') {
-                        if(form_object[v.name].indexOf(v.value) == -1){
-                            form_object[v.name].push(v.value);
-                        }
-                    } else {
-                        form_object[v.name] = [form_object[v.name]];
-                        form_object[v.name].push(v.value);
-                    }
-                } else {
-                    if(v.name == "amenity_ids" || v.name == "segmentation_ids"){
-                        form_object[v.name] = [v.value];
-                    }else{
-                        form_object[v.name] = v.value;
-                    }
-                }
-            }
-        });
-        return form_object;
-    }
-
-    /* reservation form */
-    $("#button_reservation_modal").on("click", function (e) {
-
-        if (document.documentElement.lang === "es-ES") {
-            $('input[name="new_reservation_date_modal_reservation"]').daterangepicker(
-                {
-                    locale: {
-                        direction: "ltr",
-                        format: "DD/MM/YYYY",
-                        separator: " - ",
-                        applyLabel: "Aplicar",
-                        cancelLabel: "Cancelar",
-                    },
-                    opens: "left",
-                    showCustomRangeLabel: false,
-                },
-                function (start, end, label) {
-                    console.log("cargo el calendario",label);
-                    const start_date = new Date(start);
-                    var checkin_date = start_date.toLocaleDateString(
-                        document.documentElement.lang,
-                        date_options
-                    );
-                    const end_date = new Date(end);
-                    var checkout_date = end_date.toLocaleDateString(
-                        document.documentElement.lang,
-                        date_options
-                    );
-                    $('input[name="checkin"]').val(checkin_date);
-                    $('input[name="checkout"]').val(checkout_date);
-
-                    let nights = 1;
-                    // Hours*minutes*seconds*milliseconds
-                    const oneDay = 24 * 60 * 60 * 1000;
-                    const firstDate = new Date(start);
-                    const secondDate = new Date(end);
-                    const diffDays = Math.round(
-                        Math.abs((firstDate - secondDate) / oneDay)
-                    );
-                    nights = diffDays - 1;
-                    $('input[name="nights"]').val(nights);
-                    // $("form#reservation_detail").submit();
-                }
-            );
-        } else {
-            $('input[name="new_reservation_date_modal_reservation"]').daterangepicker(
-                {
-                    locale: {
-                        direction: "ltr",
-                        format: "MM/DD/YYYY",
-                        separator: " - ",
-                    },
-                    opens: "left",
-                    showCustomRangeLabel: false,
-                },
-                function (start, end, label) {
-                    console.log("cargo el calendario",label);
-                    const start_date = new Date(start);
-                    var checkin_date = start_date.toLocaleDateString(
-                        document.documentElement.lang,
-                        date_options
-                    );
-                    const end_date = new Date(end);
-                    var checkout_date = end_date.toLocaleDateString(
-                        document.documentElement.lang,
-                        date_options
-                    );
-                    $('input[name="checkin"]').val(checkin_date);
-                    $('input[name="checkout"]').val(checkout_date);
-                    let nights = 1;
-                    // Hours*minutes*seconds*milliseconds
-                    const oneDay = 24 * 60 * 60 * 1000;
-                    const firstDate = new Date(start);
-                    const secondDate = new Date(end);
-                    const diffDays = Math.round(
-                        Math.abs((firstDate - secondDate) / oneDay)
-                    );
-                    nights = diffDays - 1;
-                    $('input[name="nights"]').val(nights);
-                    // $("form#reservation_detail").submit();
-                }
-            );
-        }
-        setTimeout(function () {
-            const today = new Date();
-            const tomorrow = new Date(today);
-            tomorrow.setDate(tomorrow.getDate() + 1);
-            var checkin_date = today.toLocaleDateString(
-                document.documentElement.lang,
-                date_options
-            );
-            var checkout_date = tomorrow.toLocaleDateString(
-                document.documentElement.lang,
-                date_options
-            );
-            $("#o_pms_pwa_new_reservation_modal")
-                .find("input[name='checkin']")
-                .val(checkin_date);
-            $("#o_pms_pwa_new_reservation_modal")
-                .find("input[name='checkout']")
-                .val(checkout_date);
-            $("#o_pms_pwa_new_reservation_modal")
-                .find("input[name='new_reservation_date_modal_reservation']")
-                .val(checkin_date + " - " + checkout_date);
-            $("#o_pms_pwa_new_reservation_modal")
-                .find("input[name='new_reservation_date_modal_reservation']")
-                .trigger("change");
-        }, 100);
-    });
-
-    $("form#booking_engine_form").on(
-        "change",
-        "input[name='new_reservation_date_modal_reservation'] ,select.call_booking_engine",
-        function (event) {
-            var values = []
-            var values = $("form#booking_engine_form").serializeArray();
-            var send_value = {};
-            var recibed_rooms = [];
-            var allowed_fields = [
-                "allowed_agency_ids",
-                "allowed_board_services",
-                // "reservation_types",
-                "allowed_channel_type_ids",
-                "allowed_pricelists",
-                "allowed_segmentations",
-                "allowed_sale_category_ids",
-                "allowed_amenity_ids",
-                // "room_types",
-                // "room_numbers",
-            ];
-            $.each(
-                values,
-                function (key,val) {
-                    if(String(val['name']).search("rooms") == 0){
-                        var rooms_array = String(val['name']).split("-");
-                        rooms_array = rooms_array.slice(1);
-                        if(!recibed_rooms[String(rooms_array[0])]){
-                            recibed_rooms[String(rooms_array[0])] = {}
-                        }
-                        if(!recibed_rooms[String(rooms_array[0])][String(rooms_array[1])]){
-                            recibed_rooms[String(rooms_array[0])][String(rooms_array[1])] = {};
-                        }
-                        recibed_rooms[String(rooms_array[0])][String(rooms_array[1])][String(rooms_array[2])] = val['value'];
-                    }
-            });
-            values = form_to_json(values);
-            if (event.currentTarget.name == "new_reservation_date_modal_reservation") {
-                // let value_range_picker = event.currentTarget.value;
-                values.checkin = $('input[name="checkin"]').val();
-                values.checkout = $('input[name="checkout"]').val();
-            }
-            var index = 0;
-            var count_rooms = 0;
-            if(recibed_rooms){
-                var send_rooms = [];
-
-                $.each(recibed_rooms, function (room_key,room_val) {
-                    if(room_val){
-                        $.each(room_val, function(key,val){
-                            if(val){
-                                send_rooms.push({
-                                    board_service_room_id: false,
-                                    checkin: values["checkin"],
-                                    checkout: values["checkout"],
-                                    preferred_room_id: val["preferred_room_id"],
-                                    pricelist_id: values["pricelist_id"],
-                                    price_per_room: val["price_per_room"],
-                                    adults: val["adults"],
-                                    room_type_id: val["room_type_id"],
-                                });
-                                count_rooms = count_rooms + 1;
-                            };
-                        });
-                        index = index + 1;
-                    };
-                });
-            }
-            send_value = {
-                partner_id: "",
-                partner_name: values["name"] || values['partner_name'],
-                mobile: values["mobile"] || "",
-                email: values["mail"] || "",
-                checkin: values["checkin"],
-                checkout: values["checkout"],
-                count_rooms_selected: count_rooms,
-                pms_property_id: values["pms_property_id"] || "1",
-                pricelist_id: values["pricelist_id"] || "1",
-                reservation_type: values["reservation_type"] || "normal",
-                rooms: send_rooms || "",
-                agrupation_type: values['agrupation_type'] || "all",
-                amenity_ids: values['amenity_ids'] || "",
-                channel_type_id: values['channel_type_id'] || "",
-
-            };
-            if (
-                values["ubication_id"] &&
-                values["ubication_id"] != "false" &&
-                values["ubication_id"] != "undefined"
-            ) {
-                send_value.ubication_id = values["ubication_id"];
-            }
-            if (
-                values["room_type_id"] &&
-                values["room_type_id"] != "false" &&
-                values["room_type_id"] != "undefined"
-            ) {
-                send_value.room_type_id = values["room_type_id"];
-            }
-            console.log(values["sale_category_id"]);
-            console.log(values['agrupation_type']);
-            if (
-                values["sale_category_id"] &&
-                values["sale_category_id"] != "false" &&
-                values["sale_category_id"] != "undefined" &&
-                values['agrupation_type'] != "room_type"
-            ) {
-                send_value.sale_category_id = values["sale_category_id"];
-            } else {
-                if(values['agrupation_type'] != "room_type"){
-                    send_value.sale_category_id = "1"; //revisar
-                }else{
-                    send_value.sale_category_id = "";
-                }
-            }
-            if (
-                values["board_service_room_id"] &&
-                values["board_service_room_id"] != "false" &&
-                values["board_service_room_id"] != "undefined"
-            ) {
-                send_value.board_service_room_id = event.getAttribute(
-                    "board_service_room_id"
-                );
-            }
-
-            if (
-                ($("#o_pms_pwa_new_reservation_modal").data("bs.modal") || {})._isShown
-            ) {
-
-                console.log("envio de inicio a booking_engine-->", send_value);
-                ajax.jsonRpc("/booking_engine", "call", send_value).then(function (
-                    new_data
-                ) {
-                    console.log("Recibo --->", new_data);
-                    setTimeout(function () {
-                        if (new_data && new_data.result != "error") {
-                            if (new_data["agrupation_type"] == "room_type") {
-                                $(".sale_category_id").removeAttr("style").hide();
-                            } else {
-                                $(".sale_category_id").show();
-                            }
-                            $.each(allowed_fields, function (key, value) {
-                                try {
-                                    var select = $(
-                                        'form#booking_engine_form [data-select="' +
-                                            value +
-                                            '"]'
-                                    );
-                                } catch (error) {
-                                    console.log(error);
-                                }
-                                if(value == "allowed_amenity_ids" || value == "allowed_segmentations"){
-                                    $.each(
-                                        new_data[value],
-                                        function (subkey, subvalue) {
-                                            if (
-                                                subvalue["id"] ==
-                                                new_data[relation_values[value]]
-                                            ) {
-                                                var option = new Option(
-                                                    subvalue["name"],
-                                                    subvalue["id"],
-                                                    true,
-                                                    true
-                                                );
-                                            } else {
-                                                var option = new Option(
-                                                    subvalue["name"],
-                                                    subvalue["id"],
-                                                    false,
-                                                    false
-                                                );
-                                            }
-                                            $(option).html(subvalue["name"]);
-                                            select.append(option);
-                                        }
-                                    );
-                                }else{
-
-                                    if (select.length != 0) {
-                                        select.empty();
-                                        if (
-                                            !new_data[relation_values[value]] &
-                                            (new_data[relation_values[value]] == 0)
-                                        ) {
-                                            select.append(
-                                                '<option value="" selected></option>'
-                                            );
-                                        }
-                                        $.each(
-                                            new_data[value],
-                                            function (subkey, subvalue) {
-
-                                                try{
-                                                    if (
-                                                        subvalue["id"] ==
-                                                        new_data[relation_values[value]].id
-                                                    ) {
-                                                        var option = new Option(
-                                                            subvalue["name"],
-                                                            subvalue["id"],
-                                                            false,
-                                                            true
-                                                        );
-                                                    } else {
-                                                        var option = new Option(
-                                                            subvalue["name"],
-                                                            subvalue["id"],
-                                                            false,
-                                                            false
-                                                        );
-                                                    }
-                                                    $(option).html(subvalue["name"]);
-                                                    select.append(option);
-                                                }catch(error) {
-                                                    //console.log(error);
-                                                    return false;
-                                                }
-                                            }
-                                        );
-                                    }
-                                }
-                                delete new_data[value];
-                            });
-                            $.each(new_data, function (key, value) {
-                                var input = $(
-                                    "form#booking_engine_form input[name='" + key + "']"
-                                );
-                                if (input.length > 0) {
-                                    input.val(value);
-                                } else {
-                                    if (!fields_to_avoid.includes(key)) {
-                                        $(
-                                            "form.o_pms_pwa_reservation_form select[name='" +
-                                                key +
-                                                "'] option[value='" +
-                                                value +
-                                                "']"
-                                        ).prop("selected", true);
-                                    }
-                                }
-                            });
-
-                            if (new_data.groups) {
-                                var groups = new_data.groups;
-                                var html = "";
-
-                                for (const i in groups) {
-                                    html +=
-                                        "<tr>" +
-                                        '<td class="col-sm-4">' +
-                                        groups[i].name +
-                                        " (" +
-                                        groups[i].max_rooms +
-                                        ")" +
-                                        "</td>" +
-                                        '<td class="col-sm-5">' +
-                                        '<button class="btn btn-o_pms_pwa_min_max form_booking_engine_group" data-id="' +
-                                        groups[i].group_id +
-                                        '" data-checkin="' +
-                                        new_data["checkin"] +
-                                        '" data-checkout="' +
-                                        new_data["checkout"] +
-                                        '" data-count_rooms_selected="' +
-                                        groups[i].count_rooms_selected +
-                                        '" data-ubication_id="' +
-                                        groups[i].ubication_id +
-                                        '" data-room_type_id="' +
-                                        groups[i].room_type_id +
-                                        '" data-sale_category_id="' +
-                                        new_data["sale_category_id"] +
-                                        '" data-pms_property_id="' +
-                                        new_data["pms_property_id"] +
-                                        '" data-pricelist_id="' +
-                                        new_data["pricelist_id"] +
-                                        '" data-board_service_room_id="' +
-                                        new_data["board_service_room_id"] +
-                                        '" onclick="document.getElementById(\'groupquantity' + groups[i].group_id +'\').stepDown(1)" data-target="#collapseme' +
-                                        groups[i].group_id +
-                                        '">-</button>' +
-                                        '<input readonly="readonly" name="count_rooms_selected"  value="' + groups[i].count_rooms_selected +'" class="o_pms_pwa_num-control" id="groupquantity' +
-                                        groups[i].group_id +
-                                        '" type="number" min="0" max="' +
-                                        groups[i].max_rooms +
-                                        '" />' +
-                                        '<button class="btn btn-o_pms_pwa_min_max form_booking_engine_group" data-id="' +
-                                        groups[i].group_id +
-                                        '" data-checkin="' +
-                                        new_data["checkin"] +
-                                        '" data-checkout="' +
-                                        new_data["checkout"] +
-                                        '" data-count_rooms_selected="' +
-                                        groups[i].count_rooms_selected +
-                                        '" data-ubication_id="' +
-                                        groups[i].ubication_id +
-                                        '" data-room_type_id="' +
-                                        groups[i].room_type_id +
-                                        '" data-sale_category_id="' +
-                                        new_data["sale_category_id"] +
-                                        '" data-pms_property_id="' +
-                                        new_data["pms_property_id"] +
-                                        '" data-pricelist_id="' +
-                                        new_data["pricelist_id"] +
-                                        '" data-board_service_room_id="' +
-                                        new_data["board_service_room_id"] +
-                                        '" onclick="document.getElementById(\'groupquantity' + groups[i].group_id +'\').stepUp(1)" data-target="#collapseme' +
-                                        groups[i].group_id +
-                                        '">+</button>' +
-                                        "</td>" +
-                                        '<td class="col-sm-3">' +
-                                        '<span class="price_group" id="price' +
-                                        i +
-                                        '">' +
-                                        groups[i].price_per_group +
-                                        "€<span>" +
-                                        "</td>" +
-                                        "</tr>" +
-                                        '<tr id="collapseme' +
-                                        groups[i].group_id +
-                                        '" class="collapse out"><td colspan="3" id="new_values_data' +
-                                        groups[i].group_id +
-                                        '"></td></tr>';
-                                }
-                                $("#bookengine_table").html(html);
-                            }
-
-                            function launchChanges(send_value) {
-                                var price_per_group = "#price" + send_value["id"];
-                                var new_values = "#new_values_data" + send_value["id"];
-                                console.log("envio a  booking_engine_group ->", send_value);
-                                ajax.jsonRpc(
-                                    "/booking_engine_group",
-                                    "call",
-                                    send_value
-                                ).then(function (new_data) {
-                                    console.log("devuelve->", new_data);
-                                    // if (new_data && new_data.result != "error") {
-                                    var html = "<table style='margin-left:-1rem; width: 100%;'>";
-
-                                    for (var i = 0; i < new_data["rooms"].length; i++) {
-                                        var seloption =
-                                            '<option name="preferred_room_id" value="' +
-                                            new_data["rooms"][i]["preferred_room_id"][
-                                                "id"
-                                            ] +
-                                            '" selected="selected">' +
-                                            new_data["rooms"][i]["preferred_room_id"][
-                                                "name"
-                                            ] +
-                                            "</option>";
-                                        $.each(
-                                            new_data["free_rooms_dict"],
-                                            function (key) {
-                                                seloption +=
-                                                    '<option  name="preferred_room_id" value="' +
-                                                    new_data["free_rooms_dict"][key][
-                                                        "id"
-                                                    ] +
-                                                    '">' +
-                                                    new_data["free_rooms_dict"][key][
-                                                        "name"
-                                                    ] +
-                                                    "</option>";
-                                            }
-                                        );
-                                        html +=
-                                            '<tr><td class="col-sm-7">' +
-                                            '<label class="control-label" for="preferred_room_id">Habitación</label>' +
-                                            '<select data-parent_id="' +
-                                            send_value["id"] +
-                                            '" data-checkin="' +
-                                            send_value["checkin"] +
-                                            '" data-checkout="' +
-                                            send_value["checkout"] +
-                                            '" data-count_rooms_selected="' +
-                                            send_value["count_rooms_selected"] +
-                                            '" data-ubication_id="' +
-                                            new_data["rooms"][i]["ubication_id"] +
-                                            '" data-room_type_id="' +
-                                            new_data["rooms"][i]["room_type_id"] +
-                                            '" data-sale_category_id="' +
-                                            send_value["sale_category_id"] +
-                                            '" data-pms_property_id="' +
-                                            send_value["pms_property_id"] +
-                                            '" data-pricelist_id="' +
-                                            send_value["pricelist_id"] +
-                                            '" data-board_service_room_id="' +
-                                            send_value["board_service_room_id"] +
-                                            '" class="form-control o_website_form_input o_domain_leaf_operator_select o_input call_booking_engine_group" name="rooms-'+send_value["id"]+'-' +
-                                            i +
-                                            '-preferred_room_id">' +
-                                            seloption +
-                                            "</select>" +
-                                            "</td>" +
-                                            '<td class="col-sm-5 text-right align-middle">' +
-                                            //'<label class="control-label" for="adults">Adultos</label>'+
-                                            '<a href="#" class="btn btn-o_pms_pwa_min_max" onclick="document.getElementById(\'quantity' + i +'\').stepDown(1)">-</a>' +
-                                            '<input name="rooms-'+send_value["id"]+'-'+ i +'-adults" class="o_pms_pwa_num-control" value="' +new_data["rooms"][i]["adults"] +'" id="quantity' +i +'" type="number" min="1" max="' + new_data["rooms"][i]["max_adults"] +'" readonly="readonly" />' +
-                                            '<input name="rooms-'+send_value["id"]+'-' +
-                                            i +
-                                            '-room_type_id" value="' +
-                                            new_data["rooms"][i]["room_type_id"] +
-                                            '" type="hidden" />' +
-                                            '<input name="rooms-'+send_value["id"]+'-' +
-                                            i +
-                                            '-price_per_room" value="' +
-                                            new_data["rooms"][i]["price_per_room"] +
-                                            '" type="hidden"/>' +
-                                            '<a href="#" class="btn btn-o_pms_pwa_min_max" onclick="document.getElementById(\'quantity' +i +'\').stepUp(1)">+</a>' +
-                                            "</td></tr>";
-                                    }
-
-                                    html += "</table>";
-                                    $(new_values).html(html);
-                                    // precio grupo
-                                    $(price_per_group).html(
-                                        new_data["price_per_group"].toFixed(2) + "€"
-                                    );
-                                    //precio total grupos
-                                    var total_price = 0.0
-                                    var price_groups_elements = document.getElementsByClassName('price_group');
-                                    for (var i = 0; i < price_groups_elements.length; ++i) {
-                                        var item = price_groups_elements[i];
-                                        total_price = (parseFloat(total_price) + parseFloat(item.innerText.replace("€", ""))).toFixed(2);
-                                    }
-                                    $("form#booking_engine_form .price_total").html(
-                                        parseFloat(total_price).toFixed(2) + "€"
-                                    );
-
-                                    // $("form#booking_engine_form .price_taxes").html(
-                                    //     (parseFloat(total_price) * 0.1).toFixed(2) + "€"
-                                    // );
-                                });
-                            }
-                            // TODO Me devuelve a 0 el número de habitaciones seleccionadas
-                            if(send_value['rooms'].length > 0){
-                                // console.log("tenemos habitaciones seleccionadas", send_value['rooms']);
-                                if (new_data.groups) {
-                                    var groups = new_data.groups;
-
-                                    for (const i in groups) {
-
-                                        if(groups[i].count_rooms_selected > 0){
-
-                                            var new_values = "#new_values_data" + groups[i]['group_id'];
-                                            var html = "<table style='margin-left:-1rem; width: 100%;'>";
-                                            for (var a = 0; a < groups[i]['rooms'].length; a++) {
-
-                                                var seloption =
-                                                    '<option name="preferred_room_id" value="' +
-                                                    groups[i]["rooms"][a][
-                                                        "preferred_room_id"
-                                                    ]["id"] +
-                                                    '" selected="selected">' +
-                                                    groups[i]["rooms"][a][
-                                                        "preferred_room_id"
-                                                    ][
-                                                        "name"
-                                                    ] +
-                                                    "</option>";
-                                                if( groups[i]["free_room_ids"].length > 0){
-                                                    $.each(
-                                                        groups[i]["free_room_ids"],
-                                                        function (key) {
-                                                            seloption +=
-                                                                '<option  name="preferred_room_id" value="' +
-                                                                groups[i]["free_room_ids"][key][
-                                                                    "id"
-                                                                ] +
-                                                                '">' +
-                                                                groups[i]["free_room_ids"][key][
-                                                                    "name"
-                                                                ] +
-                                                                "</option>";
-                                                        }
-                                                    );
-                                                }
-                                                html +=
-                                                    '<tr><td class="col-sm-7">' +
-                                                    '<label class="control-label" for="preferred_room_id">Habitación</label>' +
-                                                    '<select data-parent_id="' +
-                                                    groups[i]['group_id'] +
-                                                    '" data-checkin="' +
-                                                    send_value["checkin"] +
-                                                    '" data-checkout="' +
-                                                    send_value["checkout"] +
-                                                    '" data-count_rooms_selected="' +
-                                                    send_value["count_rooms_selected"] +
-                                                    '" data-ubication_id="' +
-                                                    groups[i]["rooms"][a]["ubication_id"] +
-                                                    '" data-room_type_id="' +
-                                                    groups[i]["rooms"][a]["room_type_id"] +
-                                                    '" data-sale_category_id="' +
-                                                    send_value["sale_category_id"] +
-                                                    '" data-pms_property_id="' +
-                                                    send_value["pms_property_id"] +
-                                                    '" data-pricelist_id="' +
-                                                    send_value["pricelist_id"] +
-                                                    '" data-board_service_room_id="' +
-                                                    send_value["board_service_room_id"] +
-                                                    '" class="form-control o_website_form_input o_domain_leaf_operator_select o_input call_booking_engine_group" name="rooms-'+groups[i]['group_id']+'-' +
-                                                    a +
-                                                    '-preferred_room_id">' +
-                                                    seloption +
-                                                    "</select>" +
-                                                    "</td>" +
-                                                    '<td class="col-sm-5 text-right align-middle">' +
-                                                    //'<label class="control-label" for="adults">Adultos</label>'+
-                                                    '<a href="#" class="btn btn-o_pms_pwa_min_max" onclick="document.getElementById(\'quantity' + a +'\').stepDown(1)">-</a>' +
-                                                    '<input name="rooms-'+groups[i]['group_id']+'-'+ a +'-adults" class="o_pms_pwa_num-control" value="' +groups[i]["rooms"][a]["adults"] +'" id="quantity' + a +'" type="number" min="1" max="' + groups[i]["rooms"][a]["max_adults"] +'" readonly="readonly" />' +
-                                                    '<input name="rooms-'+groups[i]['group_id']+'-' +
-                                                    a +
-                                                    '-room_type_id" value="' +
-                                                    groups[i]["rooms"][a]["room_type_id"] +
-                                                    '" type="hidden" />' +
-                                                    '<input name="rooms-'+groups[i]['group_id']+'-' +
-                                                    a +
-                                                    '-price_per_room" value="' +
-                                                    groups[i]["rooms"][a]["price_per_room"] +
-                                                    '" type="hidden"/>' +
-                                                    '<a href="#" class="btn btn-o_pms_pwa_min_max" onclick="document.getElementById(\'quantity' + a +'\').stepUp(1)">+</a>' +
-                                                    "</td></tr>";
-                                            }
-                                            html += "</table>";
-                                            $(new_values).html(html);
-                                            var collapseme = "#collapseme"+groups[i].group_id;
-                                            $(collapseme).collapse("show");
-                                        }
-                                    }
-                                }
-                                //launchChanges(send_value);
-
-                            }
-                            // calendar_room // TODO PRICE PER ROOM
-                            if(values.calendar_room && values.calendar_room != false){
-                                // var calendar_inputs = values.calendar_room.split("-");
-                                // console.log("calendar_inputs -->", calendar_inputs);
-                                var rooms = [{
-                                    board_service_room_id: "",
-                                    checkin: values.checkin,
-                                    checkout: values.checkout,
-                                    preferred_room_id: values["calendar_room"],
-                                    // pms_property_id': event.getAttribute("data-pms_property_id"),
-                                    pricelist_id: values.pricelist_id || "1",
-                                    adults: "1",
-                                    price_per_room: "21",
-                                    // room_type_id:
-                                    // price_per_room: values["rooms["+event.getAttribute("data-id")+"][" + i + "][price_per_room]"],
-                                    // adults: values["rooms["+event.getAttribute("data-id")+"][" + i + "][adults]"],
-                                    // room_type_id: values["rooms["+event.getAttribute("data-id")+"][" + i + "][room_type_id]"],
-                                }];
-                                values.id = "0";
-                                values.calendar_room = "";
-                                values.rooms = rooms;
-                                values.pms_property_id = String(new_data.pms_property_id) || "1";
-                                values.count_rooms_selected = "1";
-                                values.pricelist_id = values.pricelist_id || "1";
-
-                                $("input[name='calendar_room']").val("");
-                                $("input[name='count_rooms_selected']").val("1");
-
-                                launchChanges(values);
-                                $("#collapseme0").collapse("show");
-                            }
-
-                            $("form#booking_engine_form .form_booking_engine_group").on(
-                                "click",
-                                function (e) {
-                                    e.preventDefault();
-
-                                    var event = e.currentTarget;
-                                    var send_value = {};
-                                    var send_rooms = [];
-                                    var colapse_name = "#collapseme" + event.getAttribute("data-id");
-                                    var name_input =
-                                        "#groupquantity" +
-                                        event.getAttribute("data-id");
-                                    var price_per_group =
-                                        "#price" + event.getAttribute("data-id");
-                                    var num_rooms = $(name_input).val();
-                                    var values = $(
-                                        "form#booking_engine_form"
-                                    ).serializeArray();
-                                    values = form_to_json(values);
-
-                                    if (num_rooms > 0) {
-
-                                        for (
-                                            var i = 0;
-                                            i < parseInt(num_rooms);
-                                            i++
-                                        ) {
-                                            if(values[
-                                                "rooms-"+event.getAttribute("data-id")+ "-" +
-                                                    i +
-                                                    "-preferred_room_id"
-                                            ]){
-                                                send_rooms.push({
-                                                    board_service_room_id: false,
-                                                    checkin: event.getAttribute("data-checkin"),
-                                                    checkout:
-                                                        event.getAttribute("data-checkout"),
-                                                    preferred_room_id:
-                                                        values[
-                                                            "rooms-"+event.getAttribute("data-id")+ "-" +
-                                                                i +
-                                                                "-preferred_room_id"
-                                                        ],
-                                                    //'pms_property_id': event.getAttribute("data-pms_property_id"),
-                                                    pricelist_id:
-                                                        event.getAttribute("data-pricelist_id"),
-                                                    price_per_room: values["rooms-"+event.getAttribute("data-id")+ "-" + i + "-price_per_room"],
-                                                    adults: values["rooms-"+event.getAttribute("data-id")+"-" + i + "-adults"],
-                                                    room_type_id: values["rooms-"+event.getAttribute("data-id")+"-" + i + "-room_type_id"],
-                                                });
-                                            }
-                                        };
-                                        send_value = {
-                                            id: event.getAttribute("data-id"),
-                                            checkin: event.getAttribute("data-checkin"),
-                                            checkout: event.getAttribute("data-checkout"),
-                                            count_rooms_selected: num_rooms,
-                                            pms_property_id:
-                                                event.getAttribute("data-pms_property_id"),
-                                            pricelist_id:
-                                                event.getAttribute("data-pricelist_id"),
-                                            agrupation_type: values.agrupation_type,
-                                            reservation_type: values.reservation_type,
-                                            amenity_ids: values.amenity_ids || "",
-                                            channel_type_id: values.channel_type_id || "",
-                                        };
-                                        if (
-                                            event.getAttribute("data-ubication_id") &&
-                                            event.getAttribute("data-ubication_id") !=
-                                                "false" &&
-                                            event.getAttribute("data-ubication_id") !=
-                                                "undefined"
-                                        ) {
-                                            send_value.ubication_id =
-                                                event.getAttribute("data-ubication_id");
-                                        }
-                                        if (
-                                            event.getAttribute("data-room_type_id") &&
-                                            event.getAttribute("data-room_type_id") !=
-                                                "false" &&
-                                            event.getAttribute("data-room_type_id") !=
-                                                "undefined"
-                                        ) {
-                                            send_value.room_type_id =
-                                                event.getAttribute("data-room_type_id");
-                                        }
-                                        console.log(" sale_category_id 22 -> ", values.sale_category_id);
-                                        console.log(" agrupation_type 22 -> ", values.agrupation_type);
-                                        if (
-                                            values.sale_category_id &&
-                                            values.sale_category_id !=
-                                                "false" &&
-                                            values.sale_category_id !=
-                                                "undefined" &&
-                                            values.agrupation_type != "room_type"
-                                        ) {
-                                            send_value.sale_category_id =
-                                                values.sale_category_id;
-                                        } else {
-                                            if(values.agrupation_type != "room_type"){
-                                                values.sale_category_id = "1"; //revisar
-                                            }else{
-                                                values.sale_category_id = "";
-                                            }
-                                        }
-                                        if (
-                                            event.getAttribute("board_service_room_id") &&
-                                            event.getAttribute("board_service_room_id") !=
-                                                "false" &&
-                                            event.getAttribute("board_service_room_id") !=
-                                                "undefined"
-                                        ) {
-                                            send_value.board_service_room_id =
-                                                event.getAttribute("board_service_room_id");
-                                        }
-
-                                        if(send_rooms){
-                                            send_value.rooms = send_rooms
-                                        }
-                                        launchChanges(send_value);
-                                        $(colapse_name).collapse("show");
-                                    } else {
-                                        $(colapse_name).collapse("hide");
-                                        $(price_per_group).html("0 €");
-                                        var total_price = 0.0
-                                        var price_groups_elements = document.getElementsByClassName('price_group');
-                                        for (var i = 0; i < price_groups_elements.length; ++i) {
-                                            var item = price_groups_elements[i];
-                                            total_price = (parseFloat(total_price) + parseFloat(item.innerText.replace("€", ""))).toFixed(2);
-                                        }
-                                        $("form#booking_engine_form .price_total").html(
-                                            parseFloat(total_price).toFixed(2) + "€"
-                                        );
-                                        // $("form#booking_engine_form .price_taxes").html(
-                                        //     (parseFloat(total_price) * 0.1).toFixed(2) + "€"
-                                        // );
-                                    }
-                                }
-                            );
-                            $("form#booking_engine_form #bookengine_table .call_booking_engine_group").on(
-                                "change",
-                                "select",
-                                function (e) {
-
-                                    e.preventDefault();
-                                    var event = e.currentTarget;
-                                    var send_value = {};
-                                    var send_rooms = [];
-                                    var values = $(
-                                        "form#booking_engine_form"
-                                    ).serializeArray();
-                                    var name_input =
-                                    "#groupquantity" +
-                                    event.getAttribute("data-parent_id");
-                                    var num_rooms = $(name_input).val();
-                                    values = form_to_json(values);
-
-                                    for (
-                                        var i = 0;
-                                        i < parseInt(num_rooms);
-                                        i++
-                                    ) {
-                                        send_rooms.push({
-                                            board_service_room_id: false,
-                                            checkin: event.getAttribute("data-checkin"),
-                                            checkout:
-                                                event.getAttribute("data-checkout"),
-                                            preferred_room_id:
-                                                values[
-                                                    "rooms-"+event.getAttribute("data-parent_id")+"-" +
-                                                        i +
-                                                        "-preferred_room_id"
-                                                ],
-                                            //'pms_property_id': event.getAttribute("data-pms_property_id"),
-                                            pricelist_id:
-                                                event.getAttribute("data-pricelist_id"),
-                                            price_per_room: values["rooms-"+event.getAttribute("data-parent_id")+ "-" + i + "-price_per_room"],
-                                            adults: values["rooms-"+event.getAttribute("data-parent_id")+ "-" + i + "-adults"],
-                                            room_type_id: values["rooms-"+event.getAttribute("data-parent_id") + "-" + i + "-room_type_id"],
-                                        });
-                                    }
-
-                                    send_value = {
-                                        id: event.getAttribute("data-parent_id"),
-                                        checkin: event.getAttribute("data-checkin"),
-                                        checkout: event.getAttribute("data-checkout"),
-                                        count_rooms_selected: num_rooms,
-                                        pms_property_id:
-                                            event.getAttribute("data-pms_property_id"),
-                                        pricelist_id:
-                                            event.getAttribute("data-pricelist_id"),
-                                        rooms: send_rooms,
-                                        agrupation_type: values.agrupation_type,
-                                        reservation_type: values.reservation_type,
-                                        amenity_ids: values.amenity_ids || "",
-                                        channel_type_id: values.channel_type_id || "",
-                                    };
-                                    if (
-                                        event.getAttribute("data-ubication_id") &&
-                                        event.getAttribute("data-ubication_id") !=
-                                            "false" &&
-                                        event.getAttribute("data-ubication_id") !=
-                                            "undefined"
-                                    ) {
-                                        send_value.ubication_id =
-                                            event.getAttribute("data-ubication_id");
-                                    }
-                                    if (
-                                        event.getAttribute("data-room_type_id") &&
-                                        event.getAttribute("data-room_type_id") !=
-                                            "false" &&
-                                        event.getAttribute("data-room_type_id") !=
-                                            "undefined"
-                                    ) {
-                                        send_value.room_type_id =
-                                            event.getAttribute("data-room_type_id");
-                                    }
-                                    console.log(" sale_category_id 21 -> ", values.sale_category_id);
-                                    console.log(" agrupation_type 21 -> ", values.agrupation_type);
-                                    if (
-                                        values.sale_category_id &&
-                                        values.sale_category_id !=
-                                            "false" &&
-                                        values.sale_category_id !=
-                                            "undefined" &&
-                                        values.agrupation_type != "room_type"
-                                    ) {
-                                        send_value.sale_category_id =
-                                            values.sale_category_id;
-                                    } else {
-                                        if(values.agrupation_type != "room_type"){
-                                            send_value.sale_category_id = "1"; //revisar
-                                        }else{
-                                            send_value.sale_category_id = "";
-                                        }
-                                    }
-                                    if (
-                                        event.getAttribute("board_service_room_id") &&
-                                        event.getAttribute("board_service_room_id") !=
-                                            "false" &&
-                                        event.getAttribute("board_service_room_id") !=
-                                            "undefined"
-                                    ) {
-                                        send_value.board_service_room_id =
-                                            event.getAttribute("board_service_room_id");
-                                    }
-
-                                    launchChanges(send_value);
-                                }
-                            );
-                            $(
-                                "div#o_pms_pwa_new_reservation_modal #segmentation_ids"
-                            ).select2("destroy");
-                            $(
-                                "div#o_pms_pwa_new_reservation_modal #segmentation_ids"
-                            ).select2();
-                            // enviar amenity_ids
-                            $(
-                                "div#o_pms_pwa_new_reservation_modal #amenity_ids"
-                            ).select2("destroy");
-                            $(
-                                "div#o_pms_pwa_new_reservation_modal #amenity_ids"
-                            ).select2();
-                        } else {
-                            new_data.type = "warning";
-                            var alert_div = $(".o_pms_pwa_roomdoo_alerts");
-                            var alert = core.qweb.render("pms_pwa.reservation_alerts", {
-                                alert: new_data,
-                            });
-                            alert_div.append(alert);
-                        }
-                    });
-                });
-            }
-        }
-    );
-    // Envío el formulario
-    $("form#booking_engine_form #o_pms_pwa_send_form").on("click", function (e) {
-        e.preventDefault();
-        var event = e.currentTarget;
-        var send_value = {};
-        var send_rooms = [];
-        var recibed_rooms = [];
-        var values = $("form#booking_engine_form").serializeArray();
-
-        $.each(
-            values,
-            function (key,val) {
-
-                if(String(val['name']).search("rooms") == 0){
-                    var rooms_array = String(val['name']).split("-");
-                    rooms_array = rooms_array.slice(1);
-                    if(!recibed_rooms[String(rooms_array[0])]){
-                        recibed_rooms[String(rooms_array[0])] = {}
-                    }
-                    if(!recibed_rooms[String(rooms_array[0])][String(rooms_array[1])]){
-                        recibed_rooms[String(rooms_array[0])][String(rooms_array[1])] = {};
-                    }
-                    recibed_rooms[String(rooms_array[0])][String(rooms_array[1])][String(rooms_array[2])] = val['value'];
-                }
-        });
-        values = form_to_json(values);
-        var index = 0;
-        var count_rooms = 0;
-        $.each(recibed_rooms, function (room_key,room_val) {
-            if(room_val){
-                $.each(room_val, function(key,val){
-                    if(val){
-                        send_rooms.push({
-                            board_service_room_id: false,
-                            checkin: values["checkin"],
-                            checkout: values["checkout"],
-                            preferred_room_id: val["preferred_room_id"],
-                            pricelist_id: values["pricelist_id"],
-                            price_per_room: val["price_per_room"],
-                            adults: val["adults"],
-                            room_type_id: val["room_type_id"],
-                        });
-                        count_rooms = count_rooms + 1;
-                    };
-                });
-                index = index + 1;
-            };
-        });
-        send_value = {
-            partner_id: false,
-            partner_name: values["name"] || values['partner_name'],
-            mobile: values["mobile"] || "",
-            email: values["mail"] || "",
-            checkin: values["checkin"],
-            checkout: values["checkout"],
-            count_rooms_selected: count_rooms,
-            pms_property_id: values["pms_property_id"] || "1",
-            pricelist_id: values["pricelist_id"],
-            reservation_type: values["reservation_type"],
-            rooms: send_rooms,
-            channel_type_id: values['channel_type_id'] || "",
-
-        };
-        if (
-            values["ubication_id"] &&
-            values["ubication_id"] != "false" &&
-            values["ubication_id"] != "undefined"
-        ) {
-            send_value.ubication_id = values["ubication_id"];
-        }
-        if (
-            values["room_type_id"] &&
-            values["room_type_id"] != "false" &&
-            values["room_type_id"] != "undefined"
-        ) {
-            send_value.room_type_id = values["room_type_id"];
-        }
-        console.log(" sale_category_id 23-> ", values["sale_category_id"]);
-        console.log(" agrupation_type 23 -> ", values['agrupation_type']);
-        if (
-            values["sale_category_id"] &&
-            values["sale_category_id"] != "false" &&
-            values["sale_category_id"] != "undefined" &&
-            values['agrupation_type'] != "room_type"
-        ) {
-            send_value.sale_category_id = values["sale_category_id"];
-        } else {
-            if(values['agrupation_type'] != "room_type"){
-                send_value.sale_category_id = "1"; //revisar
-            }else{
-                send_value.sale_category_id = "";
-            }
-        }
-        values.sale_category_id
-        if (
-            values["board_service_room_id"] &&
-            values["board_service_room_id"] != "false" &&
-            values["board_service_room_id"] != "undefined"
-        ) {
-            send_value.board_service_room_id = event.getAttribute(
-                "board_service_room_id"
-            );
-        }
-        console.log("envio en form tras submit -->", send_value);
-        ajax.jsonRpc(
-            "/booking_engine_submit",
-            "call",
-            send_value
-        ).then(function (new_data) {
-            console.log("Recibo form tras submit -->", new_data)
-            if (new_data) {
-
-                if (new_data.reservation_id) {
-                    // Cierra modal
-                    $(
-                        "div.o_pms_pwa_new_reservation_modal"
-                    ).modal("toggle");
-                    // abre modal
-                    try{
-                        $("<td class='launch_modal' data-id='" + new_data.reservation_id + "'>Pincha aqui</td>").appendTo("table.launch_modal");
-                        var selector = "td[data-id=" +new_data.reservation_id +"]";
-                        setTimeout(function () {
-                            $(selector).click();
-                        }, 100);
-
-                    } catch (error) {
-                        console.log(error);
-                        location.href = "/reservation/" + new_data.reservation_id;
-                    }
-                } else {
-                    new_data.type = "warning";
-                    new_data.message = _t(
-                        "An undefined error has ocurred, please try again later." + new_data
-                    );
-                    var alert_div = $(".o_pms_pwa_roomdoo_alerts");
-                    var alert = core.qweb.render("pms_pwa.reservation_alerts", {
-                        alert: new_data,
-                    });
-                    alert_div.append(alert);
-                }
-            }
-        });
-    });
-
-    $("#o_pms_pwa_new_reservation_modal").on("hidden.bs.modal", function () {
-        console.log("Estoy aqui y reseteo el form");
-        $("form#booking_engine_form")[0].reset();
-        location.reload();
-
-        // const today = new Date();
-        // const tomorrow = new Date(today);
-        // tomorrow.setDate(tomorrow.getDate() + 1);
-        // var checkin_date = today.toLocaleDateString(
-        //     document.documentElement.lang,
-        //     date_options
-        // );
-        // var checkout_date = tomorrow.toLocaleDateString(
-        //     document.documentElement.lang,
-        //     date_options
-        // );
-        // $("#o_pms_pwa_new_reservation_modal")
-        //     .find("input[name='checkin']")
-        //     .val("");
-        // $("#o_pms_pwa_new_reservation_modal")
-        //     .find("input[name='checkout']")
-        //     .val("");
-        // // $("#o_pms_pwa_new_reservation_modal")
-        // //     .find("input[name='new_reservation_date_modal_reservation']")
-        // //     .val(checkin_date + " - " + checkout_date);
-    });
 
     publicWidget.registry.ReservationTableWidget = publicWidget.Widget.extend({
         selector: "table.o_pms_pwa_reservation_list_table, #o_pms_detail_reservation, table.launch_modal",
